chore: promote chat eventsub topics out of beta
