+++ eventsub-common/src/main/java/com/github/twitch4j/eventsub/subscriptions/SubscriptionTypes.java
@@ -16,10 +16,10 @@ public class SubscriptionTypes {
 
     public final @ApiStatus.Experimental BetaChannelAdBreakBeginType BETA_CHANNEL_AD_BREAK_BEGIN;
     public final ChannelBanType CHANNEL_BAN;
-    public final @ApiStatus.Experimental BetaChannelChatClearType BETA_CHANNEL_CHAT_CLEAR;
-    public final @ApiStatus.Experimental BetaChannelClearUserMessagesType BETA_CHANNEL_CLEAR_USER_MESSAGES;
-    public final @ApiStatus.Experimental BetaChannelMessageDeleteType BETA_CHANNEL_CHAT_MESSAGE_DELETE;
-    public final @ApiStatus.Experimental BetaChannelChatNotificationType BETA_CHANNEL_CHAT_NOTIFICATION;
+    public final ChannelChatClearType CHANNEL_CHAT_CLEAR;
+    public final ChannelClearUserMessagesType CHANNEL_CLEAR_USER_MESSAGES;
+    public final ChannelMessageDeleteType CHANNEL_CHAT_MESSAGE_DELETE;
+    public final ChannelChatNotificationType CHANNEL_CHAT_NOTIFICATION;
     public final ChannelCharityDonateType CHANNEL_CHARITY_DONATE;
     public final CharityCampaignStartType CHANNEL_CHARITY_START;
     public final CharityCampaignProgressType CHANNEL_CHARITY_PROGRESS;
@@ -76,10 +76,10 @@ public class SubscriptionTypes {
             Stream.of(
                 BETA_CHANNEL_AD_BREAK_BEGIN = new BetaChannelAdBreakBeginType(),
                 CHANNEL_BAN = new ChannelBanType(),
-                BETA_CHANNEL_CHAT_CLEAR = new BetaChannelChatClearType(),
-                BETA_CHANNEL_CLEAR_USER_MESSAGES = new BetaChannelClearUserMessagesType(),
-                BETA_CHANNEL_CHAT_MESSAGE_DELETE = new BetaChannelMessageDeleteType(),
-                BETA_CHANNEL_CHAT_NOTIFICATION = new BetaChannelChatNotificationType(),
+                CHANNEL_CHAT_CLEAR = new ChannelChatClearType(),
+                CHANNEL_CLEAR_USER_MESSAGES = new ChannelClearUserMessagesType(),
+                CHANNEL_CHAT_MESSAGE_DELETE = new ChannelMessageDeleteType(),
+                CHANNEL_CHAT_NOTIFICATION = new ChannelChatNotificationType(),
                 CHANNEL_CHARITY_DONATE = new ChannelCharityDonateType(),
                 CHANNEL_CHARITY_START = new CharityCampaignStartType(),
                 CHANNEL_CHARITY_PROGRESS = new CharityCampaignProgressType(),
